requirements.txt
passlib
psutil
ollama
google-generativeai
google-genai
